Input Validation and Exception Handling (#80)
* Input Validation and Exception Handling

Input Validation:

The input data is validated to ensure it's neither null nor an empty string. If so, the response status is INVALID_REQUEST, with an appropriate error message.
It checks whether the input consists solely of digits. If the input contains any non-digit characters, the status is INVALID_REQUEST, with an error message indicating that the input is not a number.

Exception Handling:

A try-catch block is added to handle unexpected exceptions. If an exception occurs, it returns a response with an ERROR status and a message that provides information about the unexpected issue.
Specific exception handling is also implemented for cases where the input string cannot be parsed into a long, returning an appropriate error message when necessary

* Fix errors and edit comments

Fix errors in ComputeSystemImpl.java
It was an error that occurred because I accidentally deleted part of the code.

I edited old comments.