Assignment 7-2 (#111)
* Assignment 7-2

Create a basic frontend for engine

a. Create a java client file with a main method
b. Set up gRPC and protocol buffers to support your user <-> compute server API (Network API 1, the controller/coordinator/manager component)
c. Allow a user to specify (via Scanner prompts, command line arguments, or other)
i. either a file to upload or a typed-in list of numbers for computation
ii. An output file
iii. Optionally, a delimiter to use for the output file
d. Provide a way to let the user know when the task is complete, and if it succeeded or failed

* NetworkAPIServer.java uncomment